CLIM-402: Implement "custom user shapefiles" feature
Add the "custom shapefile" feature to the "download" page. When the user selects a shapefile in the "Custom shapefile" file input of the "Area" tab, load the file, validate it (make sure it contains polygons), simplify it and show it on the map. Also allow the user to select regions from the file and allow the code to extract the json data for the region.