CLIM-476 - Add a "Download" button in "Explore Maps"
The "Export Data" section of the "Download" tab is replaced by this button.

Clicking on the button uses the current page's query to build a URL for the "Download Data" and redirects to it. The built URL will allow the "Download Data" page to prefill some inputs with the values of the current page.

Still missing some official text and translation.