CLIM-715: Fix: Builder strips page content when too many blocks (#254)
The Builder strips the last parts of the page when saving a page that has more than a certain number of blocks and attributes. This is due to a limit in the number of POST parameters allowed by PHP.

Fix this bug by changing how the JavaScript builds the POST request to use only one parameter.