refs #81968: Fix typography issues in the app
examined different ways to load the fonts via `vite.config.ts` including:
- add base to the vite config to match the url where app is served
- leveraging vite\'s publicDir to load the fonts
- using relative path to load the fonts instead of alias
- directly importing the fonts in `App.ts`

None of the above worked.

The next step was to load a separate font file via vite as a
separate entry in addition to explicitly adding it to the templates via the
`cdc_app_asset_load()` function. That did not work either.

The last resort was to by pass all the build processes and load
the fonts directly in the templates from the theme\'s assets and to
be written directly in the templates head element which resolved the issue.