81992: Add necessary behaviour to the variable selection filters to refine the list of variables -- improvements (#373)
* Refs: 81992: All datasets are loading on first load.

instead of getting the list of dataset every time the dataset panel is opened.
The default variable loaded is the first variable found in that dataset.

* Refs: 81992: The variable panel can now be scrolled

when there are a lot of variables.

* Refs: 81992: correct scroll effect for the datasets panel

similar to the variables panel

* Ref: 81992: Correct scroll color for the sidebar content component

so if it scrolls just in case

* Refs: 81992: Make initialState empty from  the first variable config

which was initially done for testing purposes

* Fix: Cache variable list to prevent reloading on panel open

The variable list was reloading every time the variables panel was opened.
I had to lift up the variable data to Redux,
And by implementing caching for variables,
similar to the approach used for datasets,
but reloading if a new dataset is selected.