Move internal URLs out of TypeScript
Internal URLs, used to refer to other pages of the site, are used in some places. Currently the URLs are hardcoded in the TypeScript code.

Move the definitions of those URLs into the PHP, where logic could eventually be used to have the URLs dynamic and not hardcoded.

Also closes CLIM-1001 and CLIM-1002