CLIM-855: Simplify and improve development environment setup (#340)
The setup of the development environment was requiring multiple steps, the creation of a file, directory permission changes, etc.

This commit introduces changes to the Docker image and compose files that simplify the development setup and remove the need to create custom files and to change permissions. It also introduces a new dev.sh script containing commands to centralize the main development commands.

It also removes the dependency on NVM since it wasn't required.

The documentation was updated.